+++ backend/internal/services/eating/managers/helpers_test.go
@@ -1,0 +1,85 @@
+package managers
+
+import (
+	"slices"
+	"testing"
+
+	msgconfig "github.com/dinnerdonebetter/backend/internal/lib/messagequeue/config"
+	mockpublishers "github.com/dinnerdonebetter/backend/internal/lib/messagequeue/mock"
+	"github.com/dinnerdonebetter/backend/internal/lib/observability/logging"
+	"github.com/dinnerdonebetter/backend/internal/lib/observability/metrics"
+	"github.com/dinnerdonebetter/backend/internal/lib/observability/tracing"
+	textsearchcfg "github.com/dinnerdonebetter/backend/internal/lib/search/text/config"
+	"github.com/dinnerdonebetter/backend/internal/lib/testutils"
+	"github.com/dinnerdonebetter/backend/internal/services/eating/database"
+	"github.com/dinnerdonebetter/backend/internal/services/eating/types"
+
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/require"
+)
+
+func eventMatches(eventType string, keys []string) any {
+	return mock.MatchedBy(func(message *types.DataChangeMessage) bool {
+		allContextKeys := []string{}
+		for k := range message.Context {
+			allContextKeys = append(allContextKeys, k)
+		}
+
+		slices.Sort(keys)
+		slices.Sort(allContextKeys)
+		allKeysMatch := slices.Equal(keys, allContextKeys)
+		eventTypeMatches := message.EventType == eventType
+		result := allKeysMatch && eventTypeMatches
+
+		return result
+	})
+}
+
+func setupExpectations(
+	manager *mealPlanningManager,
+	dbSetupFunc func(db *database.MockDatabase),
+	eventTypeMaps ...map[string][]string,
+) []any {
+	db := database.NewMockDatabase()
+	if dbSetupFunc != nil {
+		dbSetupFunc(db)
+	}
+	manager.db = db
+
+	mp := &mockpublishers.Publisher{}
+	for _, eventTypeMap := range eventTypeMaps {
+		for eventType, payload := range eventTypeMap {
+			mp.On("PublishAsync", testutils.ContextMatcher, eventMatches(eventType, payload)).Return()
+		}
+	}
+	manager.dataChangesPublisher = mp
+
+	return []any{db, mp}
+}
+
+func buildMealPlanManagerForTest(t *testing.T) *mealPlanningManager {
+	t.Helper()
+
+	queueCfg := &msgconfig.QueuesConfig{
+		DataChangesTopicName: t.Name(),
+	}
+
+	mpp := &mockpublishers.PublisherProvider{}
+	mpp.On("ProvidePublisher", queueCfg.DataChangesTopicName).Return(&mockpublishers.Publisher{}, nil)
+
+	m, err := NewMealPlanningManager(
+		t.Context(),
+		logging.NewNoopLogger(),
+		tracing.NewNoopTracerProvider(),
+		database.NewMockDatabase(),
+		queueCfg,
+		mpp,
+		&textsearchcfg.Config{},
+		metrics.NewNoopMetricsProvider(),
+	)
+	require.NoError(t, err)
+
+	mock.AssertExpectationsForObjects(t, mpp)
+
+	return m.(*mealPlanningManager)
+}
